requirements.txt
asgiref==3.2.3
Django==3.0.1
django-redis==4.11.0
djangorestframework==3.11.0
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
